+++ features/support/world.rb
@@ -1,0 +1,10 @@
+module DebifyWorld
+  attr_accessor :containers, :networks
+
+  def initialize
+    @containers = []
+    @networks = []
+  end
+end
+
+World(DebifyWorld)